ADD: GetAllReservations Edit: post request body for movie projection

+++ src/main/java/ferit/cinema/feature/auditorium/AuditoriumRepository.java
@@ -2,5 +2,5 @@ package ferit.cinema.feature.auditorium;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuditoriumRepository extends JpaRepository<Auditorium, Integer> {
+public interface AuditoriumRepository extends JpaRepository<Auditorium, Long> {
 }
